+++ src/views/Attendance/SalaryProcess/MonthlySalaryProcess.jsx
@@ -5,19 +5,21 @@ import moment from 'moment';
 import { Button, Chip, Input, Sheet, Table, Typography } from '@mui/joy';
 import { DatePicker, LocalizationProvider } from '@mui/x-date-pickers';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
-import { addMonths, endOfMonth, format, getDaysInMonth, isValid, startOfMonth, subMonths } from 'date-fns';
+import { addMonths, endOfMonth, format, isValid, startOfMonth, subMonths } from 'date-fns';
 import HourglassEmptyOutlinedIcon from '@mui/icons-material/HourglassEmptyOutlined';
 import { axioslogin } from 'src/views/Axios/Axios';
-import { errorNofity, warningNofity } from 'src/views/CommonCode/Commonfunc';
+import { errorNofity, succesNofity, warningNofity } from 'src/views/CommonCode/Commonfunc';
 import CalendarMonthIcon from '@mui/icons-material/CalendarMonth';
 import { useSelector } from 'react-redux';
 import UploadIcon from '@mui/icons-material/Upload';
 import { getAllDeptList } from 'src/redux/actions/Department.action';
-import { useQuery, useQueryClient } from 'react-query';
+import { useQuery, } from 'react-query';
 import { getAllDeptSectList } from 'src/redux/actions/DepartmentSection.Action';
 import CustomBackDrop from 'src/views/Component/MuiCustomComponent/CustomBackDrop';
 import { getCommonsettingData } from 'src/redux/actions/Common.Action';
 import { ExportAttendance } from 'src/views/Payroll/AttendanceUpdation/ExportToExcel';
+import { ToastContainer } from 'react-toastify';
+import { attendnaceCountCalculationFunc, employeeEarnDeduction, getAllPunchmastData } from './SalaryProcessFunctions';
 
 const MonthlySalaryProcess = () => {
     const [value, setValue] = useState(moment(new Date()));
@@ -35,18 +37,21 @@ const MonthlySalaryProcess = () => {
 
     const { data: deptartments } = useQuery({
         queryKey: ['departmentList'],
-        queryFn: getAllDeptList
+        queryFn: getAllDeptList,
+        staleTime: Infinity
     })
 
     const { data: deptartmentSection } = useQuery({
         queryKey: ['departmentSectionList'],
-        queryFn: getAllDeptSectList
+        queryFn: getAllDeptSectList,
+        staleTime: Infinity
     })
 
 
     const { data: commonSettings } = useQuery({
         queryKey: ['commonSettingdata'],
-        queryFn: getCommonsettingData
+        queryFn: getCommonsettingData,
+        staleTime: Infinity
     })
 
     const onProcessClick = useCallback(async () => {
@@ -59,14 +64,12 @@ const MonthlySalaryProcess = () => {
         const result = await axioslogin.get('/payrollprocess/getAcriveDepartmentSection/');
         const { success, data } = result.data;
         const deptSectionData = data;
-        console.log(deptSectionData);
         if (success === 1) {
             //GET PUNCHMARKING DATA FROM table 
             const getPunchMarkingHr_table = await axioslogin.post('/payrollprocess/getmonthdeptlist/', getPunchMarkTablePostData);
             const { succ, data } = getPunchMarkingHr_table.data;
             if (succ === 1) {
                 const punchMarkingTableData = data;
-                console.log(punchMarkingTableData);
                 const findDept = [...new Set(deptSectionData?.map(e => e.dept_id))]?.map((dept) => {
                     return {
                         "dept_id": dept,
@@ -110,7 +113,6 @@ const MonthlySalaryProcess = () => {
                         }
                     })
                     setArray(array)
-                    console.log(findDept);
                     setDeptList(findDept)
 
                     setOpenBkDrop(false)
@@ -121,153 +123,252 @@ const MonthlySalaryProcess = () => {
                 setDeptList(findDept)
                 setOpenBkDrop(false)
             } else if (succ === 2) {
+                const deptArray = deptartments?.map(val => val.dept_id)
+                const sectArray = deptartmentSection?.map(val => val.sect_id)
+                const getEmpData = {
+                    em_department: deptArray,
+                    em_dept_section: sectArray,
+                }
 
-                // IF NO DATA -> INSERT IN TO payroll_processed_salary table
-                const postData_PunchMarkHR = deptSectionData?.map((e) => {
-                    return [
-                        format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
-                        e.dept_id,
-                        e.sect_id,
-                        1,
-                        em_no,
-                        em_no,
-                        format(endOfMonth(new Date(value)), 'yyyy-MM-dd')
-                    ]
-                })
-                const insertPunchMarkTable = await axioslogin.post('/payrollprocess/insert/monthlyprocess', postData_PunchMarkHR)
-                const { success } = insertPunchMarkTable.data
-                if (success === 1) {
-                    const findDept = [...new Set(deptSectionData?.map(e => e.dept_id))]?.map((dept) => {
-                        return {
-                            "dept_id": dept,
-                            "dept_name": deptSectionData?.find(e => e.dept_id === dept)?.dept_name,
-                            "section": deptSectionData?.filter((val) => val.dept_id === dept).map((v) => {
-                                return { ...v, "updated": format(startOfMonth(new Date(value)), 'yyyy-MM-dd') }
-                            }),
-                        }
-                    })
-                    const deptArray = deptartments?.map(val => val.dept_id)
-                    const sectArray = deptartmentSection?.map(val => val.sect_id)
-                    const getEmpData = {
-                        em_department: deptArray,
-                        em_dept_section: sectArray,
-                    }
-                    const result1 = await axioslogin.post("/payrollprocess/getAllEmployee", getEmpData);
-                    const { succes, dataa: employeeData } = result1.data
-                    if (succes === 1 || isValid(value) || value !== null) {
+                // const getEmpData = {
+                //     em_department: [15, 1],
+                //     em_dept_section: [13],
+                // }
 
-                        const result1 = await axioslogin.post("/payrollprocess/empDeduction", getEmpData)
-                        const { data: deductData } = result1.data
 
-                        const arr = employeeData && employeeData.map((val) => val.em_id)
-                        const postdata = {
-                            emp_id: arr,
-                            // emp_id: [168],
-                            from: format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
-                            to: format(endOfMonth(new Date(value)), 'yyyy-MM-dd'),
-                        }
-                        const result = await axioslogin.post("/payrollprocess/punchbiId", postdata);
-                        const { success, data } = result.data
-                        if (success === 1) {
-                            const finalDataArry = employeeData?.map((val) => {
-                                const empwise = data.filter((value) => value.emp_id === val.em_id)
+                const result1 = await axioslogin.post("/payrollprocess/getAllEmployee", getEmpData);
+                const { succes, dataa: employeeData } = result1.data
+                if (succes === 1 && isValid(value) && value !== null) {
 
-                                const totalH = (empwise?.filter(val => val.holiday_status === 1)).length
-                                //  const totalLOP = (empwise?.filter(val => val.lvereq_desc === 'A' || val.lvereq_desc === 'ESI' || val.lvereq_desc === 'LWP' || val.lvereq_desc === 'ML')).length
-                                const totalLV = (empwise?.filter(val => val.lvereq_desc === 'SL' || val.lvereq_desc === 'CL' || val.lvereq_desc === 'COFF' || val.lvereq_desc === 'EL')).length
-                                const totalHD = (empwise?.filter(val => val.lvereq_desc === 'HD' || val.lvereq_desc === 'CHD' || val.lvereq_desc === 'EGHD' || val.lvereq_desc === 'HDSL' || val.lvereq_desc === 'HDCL')).length
-                                const totalLC = (empwise?.filter(val => val.lvereq_desc === 'LC')).length
+                    employeeEarnDeduction(getEmpData).then((values) => {
+                        const { status, data: deductData } = values;
+                        if (status === 1) {
+                            const arr = employeeData && employeeData.map((val) => val.em_id)
+                            const postdata = {
+                                emp_id: arr,
+                                from: format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                                to: format(endOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                            }
+                            getAllPunchmastData(postdata).then((punchmastdata) => {
+                                const { status, data } = punchmastdata;
+                                if (status === 1) {
 
-                                const deductValue = (deductData?.filter(item => val.em_no === item.em_no).reduce((acc, curr) => acc + (curr.em_amount), 0)) ?? 0;
+                                    attendnaceCountCalculationFunc(employeeData, deductData, data, value, commonSettings).then(async (allData) => {
+                                        const { status, data } = allData
+                                        if (status === 1 && data?.length !== 0) {
+                                            //console.log(data)
+                                            setArray(data)
+                                            setOpenBkDrop(false)
 
-                                const npsamount = val.nps === 1 ? val.npsamount : 0
-                                const lwfamount = val.lwf_status === 1 ? val.lwfamount : 0
+                                            const result1 = await axioslogin.post("/payrollprocess/create/processedSalary", data);
+                                            const { success, message } = result1.data
+                                            if (success === 1) {
+                                                // IF NO DATA -> INSERT IN TO payroll_processed_salary table
+                                                const postData_PunchMarkHR = deptSectionData?.map((e) => {
+                                                    return [
+                                                        format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                                                        e.dept_id,
+                                                        e.sect_id,
+                                                        1,
+                                                        em_no,
+                                                        em_no,
+                                                        format(endOfMonth(new Date(value)), 'yyyy-MM-dd')
+                                                    ]
+                                                })
+                                                // if condition for check null and undefined 
+                                                const insertPunchMarkTable = await axioslogin.post('/payrollprocess/insert/monthlyprocess', postData_PunchMarkHR)
+                                                const { success } = insertPunchMarkTable.data
+                                                if (success === 1) {
+                                                    const findDept = [...new Set(deptSectionData?.map(e => e.dept_id))]?.map((dept) => {
+                                                        return {
+                                                            "dept_id": dept,
+                                                            "dept_name": deptSectionData?.find(e => e.dept_id === dept)?.dept_name,
+                                                            "section": deptSectionData?.filter((val) => val.dept_id === dept).map((v) => {
+                                                                return {
+                                                                    ...v, "updated": format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                                                                    "status": 1
+                                                                }
+                                                            }),
+                                                        }
+                                                    })
+                                                    setArray(data)
+                                                    setDeptList(findDept)
+                                                    setOpenBkDrop(false)
 
-                                const onedaySalary = val.gross_salary / getDaysInMonth(new Date(value))
+                                                } else {
+                                                    warningNofity("Error Updating the Punchmarking HR Data ! contact IT")
+                                                    setOpenBkDrop(false)
+                                                }
 
-                                // const totallopCount = totalLC > commonSettings?.max_late_day_count ? totalLOP + (totalHD * 0.5) + ((totalLC - commonSettings?.max_late_day_count) / 2) : totalLOP + (totalHD * 0.5)
-                                // const totallopCount = totalLOP + (totalHD * 0.5)
-
-                                const workday =
-                                    (empwise?.filter(val => val.lvereq_desc === 'P' || val.lvereq_desc === 'WOFF' ||
-                                        val.lvereq_desc === 'COFF' || val.lvereq_desc === 'NOFF' || val.lvereq_desc === 'DOFF' ||
-                                        val.lvereq_desc === 'SL' || val.lvereq_desc === 'HP' ||
-                                        val.lvereq_desc === 'CL' || val.lvereq_desc === 'EL' ||
-                                        val.lvereq_desc === 'H' || val.lvereq_desc === 'OHP' ||
-                                        val.lvereq_desc === 'ODP' || val.lvereq_desc === 'OBS' || val.lvereq_desc === 'LC')).length
-
-                                const totalHP = (empwise?.filter(val => val.lvereq_desc === 'HP')).length
-
-                                const totalDays = getDaysInMonth(new Date(value))
-                                const holidaysalary = val.gross_salary <= commonSettings.salary_above ? onedaySalary * totalHP : 0;
-                                const totalPayday = workday + (totalHD * 0.5)
-                                const totallopCount = totalDays - totalPayday;
-                                // const totalPayday = workday === 0 ? 0 : totalDays - totallopCount
-                                const lopamount = totallopCount * (val.gross_salary / totalDays);
-                                //const paydaySalay = (val.gross_salary / totalDays) * totalPayday
-                                const totalSalary = Number(val.gross_salary).toFixed(2) - Number(npsamount).toFixed(2) - Number(lwfamount).toFixed(2) - Number(deductValue).toFixed(2) - Number(lopamount).toFixed(2)
-
-                                return {
-                                    em_no: val.em_no,
-                                    em_name: val.em_name,
-                                    branch_name: val.branch_name,
-                                    dept_name: val.dept_name,
-                                    sect_name: val.sect_name,
-                                    ecat_name: val.ecat_name,
-                                    desg_name: val.desg_name,
-                                    inst_emp_type: val.inst_emp_type,
-                                    empSalary: val.gross_salary,
-                                    em_account_no: val.em_account_no,
-                                    em_ifsc: val.em_ifsc,
-                                    totalDays: getDaysInMonth(new Date(value)),
-                                    totalLeaves: totalLV,
-                                    totalHoliday: totalH,
-                                    totallopCount: totalPayday === 0 ? getDaysInMonth(new Date(value)) : totallopCount,
-                                    holidayworked: totalHP,
-                                    totalHD: totalHD,
-                                    totalLC: totalLC,
-                                    paydays: totalPayday,
-                                    lopAmount: Math.round(onedaySalary * totallopCount),
-                                    npsamount: npsamount,
-                                    lwfamount: lwfamount,
-                                    holidaySalary: Math.round(holidaysalary),
-                                    deductValue: deductValue,
-                                    totalSalary: totalSalary < 0 ? 0 : Math.round(totalSalary),
-                                    branch_slno: val.branch_slno,
-                                    category_slno: val.category_slno,
-                                    dept_id: val.dept_id,
-                                    desg_slno: val.desg_slno,
-                                    em_id: val.em_id,
-                                    inst_slno: val.inst_slno,
-                                    sect_id: val.sect_id,
-                                    processed_month: format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                                            } else {
+                                                errorNofity(message)
+                                                setOpenBkDrop(false)
+                                            }
+                                        } else {
+                                            errorNofity("Error While Attendance Calculation")
+                                            setOpenBkDrop(false)
+                                        }
+                                    })
+                                } else {
+                                    errorNofity("Error While Geting All Employee PunchMast Data")
+                                    setOpenBkDrop(false)
                                 }
                             })
-                            const result1 = await axioslogin.post("/payrollprocess/create/processedSalary", finalDataArry);
-                            const { success, message } = result1.data
-                            if (success === 1) {
-                                setArray(finalDataArry)
-                                setDeptList(findDept)
-
-                                setOpenBkDrop(false)
-                            } else {
-                                errorNofity(message)
-                                setOpenBkDrop(false)
-                            }
-                        }
-                        else {
-                            warningNofity("No Punch Details")
+                        } else {
+                            errorNofity("Error While Getting Deduction Details!")
                             setOpenBkDrop(false)
                         }
-                    } else {
-                        warningNofity("Error While Fetching data!")
-                        setOpenBkDrop(false)
-                    }
-                    // setOpenBkDrop(false)
+                    })
                 } else {
-                    warningNofity("Error Updating the Punchmarking HR Data ! contact IT")
-                    //setOpenBkDrop(false)
+                    warningNofity("Error While Fetching Employee data!")
+                    setOpenBkDrop(false)
                 }
+
+                // // IF NO DATA -> INSERT IN TO payroll_processed_salary table
+                // const postData_PunchMarkHR = deptSectionData?.map((e) => {
+                //     return [
+                //         format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                //         e.dept_id,
+                //         e.sect_id,
+                //         1,
+                //         em_no,
+                //         em_no,
+                //         format(endOfMonth(new Date(value)), 'yyyy-MM-dd')
+                //     ]
+                // })
+                // // if condition for check null and undefined 
+                // const insertPunchMarkTable = await axioslogin.post('/payrollprocess/insert/monthlyprocess', postData_PunchMarkHR)
+                // const { success } = insertPunchMarkTable.data
+                // if (success === 1) {
+                //     const findDept = [...new Set(deptSectionData?.map(e => e.dept_id))]?.map((dept) => {
+                //         return {
+                //             "dept_id": dept,
+                //             "dept_name": deptSectionData?.find(e => e.dept_id === dept)?.dept_name,
+                //             "section": deptSectionData?.filter((val) => val.dept_id === dept).map((v) => {
+                //                 return { ...v, "updated": format(startOfMonth(new Date(value)), 'yyyy-MM-dd') }
+                //             }),
+                //         }
+                //     })
+
+                //     // check department and department section > 0
+                //     const result1 = await axioslogin.post("/payrollprocess/getAllEmployee", getEmpData);
+
+                //     //  getEmpData.length > 0 and null and undefined 
+                //     const { succes, dataa: employeeData } = result1.data
+                //     if (succes === 1 || isValid(value) || value !== null) {
+
+                //         const result1 = await axioslogin.post("/payrollprocess/empDeduction", getEmpData)
+                //         const { data: deductData } = result1.data
+
+                //         const arr = employeeData && employeeData.map((val) => val.em_id)
+                //         const postdata = {
+                //             emp_id: arr,
+                //             // emp_id: [168],
+                //             from: format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                //             to: format(endOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                //         }
+                //         const result = await axioslogin.post("/payrollprocess/punchbiId", postdata);
+                //         const { success, data } = result.data
+                //         if (success === 1) {
+                //             const finalDataArry = employeeData?.map((val) => {
+                //                 const empwise = data.filter((value) => value.emp_id === val.em_id)
+
+                //                 const totalH = (empwise?.filter(val => val.holiday_status === 1)).length
+                //                 //  const totalLOP = (empwise?.filter(val => val.lvereq_desc === 'A' || val.lvereq_desc === 'ESI' || val.lvereq_desc === 'LWP' || val.lvereq_desc === 'ML')).length
+                //                 const totalLV = (empwise?.filter(val => val.lvereq_desc === 'SL' || val.lvereq_desc === 'CL' || val.lvereq_desc === 'COFF' || val.lvereq_desc === 'EL')).length
+                //                 const totalHD = (empwise?.filter(val => val.lvereq_desc === 'HD' || val.lvereq_desc === 'CHD' || val.lvereq_desc === 'EGHD' || val.lvereq_desc === 'HDSL' || val.lvereq_desc === 'HDCL')).length
+                //                 const totalLC = (empwise?.filter(val => val.lvereq_desc === 'LC')).length
+
+                //                 const deductValue = (deductData?.filter(item => val.em_no === item.em_no).reduce((acc, curr) => acc + (curr.em_amount), 0)) ?? 0;
+
+                //                 const npsamount = val.nps === 1 ? val.npsamount : 0
+                //                 const lwfamount = val.lwf_status === 1 ? val.lwfamount : 0
+
+                //                 const onedaySalary = val.gross_salary / getDaysInMonth(new Date(value))
+
+                //                 // const totallopCount = totalLC > commonSettings?.max_late_day_count ? totalLOP + (totalHD * 0.5) + ((totalLC - commonSettings?.max_late_day_count) / 2) : totalLOP + (totalHD * 0.5)
+                //                 // const totallopCount = totalLOP + (totalHD * 0.5)
+
+                //                 const workday =
+                //                     (empwise?.filter(val => val.lvereq_desc === 'P' || val.lvereq_desc === 'WOFF' ||
+                //                         val.lvereq_desc === 'COFF' || val.lvereq_desc === 'NOFF' || val.lvereq_desc === 'DOFF' ||
+                //                         val.lvereq_desc === 'SL' || val.lvereq_desc === 'HP' ||
+                //                         val.lvereq_desc === 'CL' || val.lvereq_desc === 'EL' ||
+                //                         val.lvereq_desc === 'H' || val.lvereq_desc === 'OHP' ||
+                //                         val.lvereq_desc === 'ODP' || val.lvereq_desc === 'OBS' || val.lvereq_desc === 'LC')).length
+
+                //                 const totalHP = (empwise?.filter(val => val.lvereq_desc === 'HP')).length
+
+                //                 const totalDays = getDaysInMonth(new Date(value))
+                //                 const holidaysalary = val.gross_salary <= commonSettings.salary_above ? onedaySalary * totalHP : 0;
+                //                 const totalPayday = workday + (totalHD * 0.5)
+                //                 const totallopCount = totalDays - totalPayday;
+                //                 // const totalPayday = workday === 0 ? 0 : totalDays - totallopCount
+                //                 const lopamount = totallopCount * (val.gross_salary / totalDays);
+                //                 //const paydaySalay = (val.gross_salary / totalDays) * totalPayday
+                //                 const totalSalary = Number(val.gross_salary).toFixed(2) - Number(npsamount).toFixed(2) - Number(lwfamount).toFixed(2) - Number(deductValue).toFixed(2) - Number(lopamount).toFixed(2)
+
+                //                 return {
+                //                     em_no: val.em_no,
+                //                     em_name: val.em_name,
+                //                     branch_name: val.branch_name,
+                //                     dept_name: val.dept_name,
+                //                     sect_name: val.sect_name,
+                //                     ecat_name: val.ecat_name,
+                //                     desg_name: val.desg_name,
+                //                     inst_emp_type: val.inst_emp_type,
+                //                     empSalary: val.gross_salary,
+                //                     em_account_no: val.em_account_no,
+                //                     em_ifsc: val.em_ifsc,
+                //                     totalDays: getDaysInMonth(new Date(value)),
+                //                     totalLeaves: totalLV,
+                //                     totalHoliday: totalH,
+                //                     totallopCount: totalPayday === 0 ? getDaysInMonth(new Date(value)) : totallopCount,
+                //                     holidayworked: totalHP,
+                //                     totalHD: totalHD,
+                //                     totalLC: totalLC,
+                //                     paydays: totalPayday,
+                //                     lopAmount: Math.round(onedaySalary * totallopCount),
+                //                     npsamount: npsamount,
+                //                     lwfamount: lwfamount,
+                //                     holidaySalary: Math.round(holidaysalary),
+                //                     deductValue: deductValue,
+                //                     totalSalary: totalSalary < 0 ? 0 : Math.round(totalSalary),
+                //                     branch_slno: val.branch_slno,
+                //                     category_slno: val.category_slno,
+                //                     dept_id: val.dept_id,
+                //                     desg_slno: val.desg_slno,
+                //                     em_id: val.em_id,
+                //                     inst_slno: val.inst_slno,
+                //                     sect_id: val.sect_id,
+                //                     processed_month: format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                //                 }
+                //             })
+                //             const result1 = await axioslogin.post("/payrollprocess/create/processedSalary", finalDataArry);
+                //             const { success, message } = result1.data
+                //             if (success === 1) {
+                //                 setArray(finalDataArry)
+                //                 setDeptList(findDept)
+
+                //                 setOpenBkDrop(false)
+                //             } else {
+                //                 errorNofity(message)
+                //                 setOpenBkDrop(false)
+                //             }
+                //         }
+                //         else {
+                //             warningNofity("No Punch Details")
+                //             setOpenBkDrop(false)
+                //         }
+                //     } else {
+                //         warningNofity("Error While Fetching data!")
+                //         setOpenBkDrop(false)
+                //     }
+                //     // setOpenBkDrop(false)
+                // } else {
+                //     warningNofity("Error Updating the Punchmarking HR Data ! contact IT")
+                //     //setOpenBkDrop(false)
+                // }
             } else {
                 // IF ERROR
                 warningNofity('----error getting punchmarking table data ! contact IT Department')
@@ -279,13 +380,17 @@ const MonthlySalaryProcess = () => {
         }
     }, [value, em_no, deptartments, deptartmentSection, commonSettings])
 
+    //  const loginem_id = useSelector((state) => state?.getProfileData?.ProfileData[0]?.em_id ?? 0)
+
     const deleteAttendanceMarkingProcess = useCallback(async (dept, section, date) => {
-        console.log(dept, section, date);
+        const getPunchMarkTablePostData = {
+            month: format(startOfMonth(new Date(date)), 'yyyy-MM-dd')
+        }
         const postDta = {
-            update_user: 4516,
+            update_user: em_no,
             dept_id: dept,
             sect_id: section,
-            processed_month: format(startOfMonth(new Date(value)), 'yyyy-MM-dd')
+            processed_month: format(startOfMonth(new Date(date)), 'yyyy-MM-dd')
         }
 
         const result1 = await axioslogin.post("/payrollprocess/delete/processedSalary", postDta);
@@ -294,17 +399,44 @@ const MonthlySalaryProcess = () => {
             const result1 = await axioslogin.post("/payrollprocess/cancel/process", postDta);
             const { success, message } = result1.data
             if (success === 1) {
-                //setDeptList(findDept)
-                setOpenBkDrop(false)
+                // setDeptList([])
+                // //setDeptList(findDept)
+                // setOpenBkDrop(false)
+                // succesNofity(message)
+                const result = await axioslogin.get('/payrollprocess/getAcriveDepartmentSection/');
+                const { success, data } = result.data;
+                const deptSectionData = data;
+                if (success === 1) {
+                    const getPunchMarkingHr_table = await axioslogin.post('/payrollprocess/getmonthdeptlist/', getPunchMarkTablePostData);
+                    const { succ, data } = getPunchMarkingHr_table.data;
+                    if (succ === 1) {
+                        const punchMarkingTableData = data;
+                        const findDept = [...new Set(deptSectionData?.map(e => e.dept_id))]?.map((dept) => {
+                            return {
+                                "dept_id": dept,
+                                "dept_name": deptSectionData?.find(e => e.dept_id === dept)?.dept_name,
+                                "section": deptSectionData?.filter((val) => val.dept_id === dept).map((v) => {
+                                    return {
+                                        ...v, "updated": punchMarkingTableData?.find((e) => v.sect_id === e.sect_id)?.last_update_date ?? format(startOfMonth(new Date(value)), 'yyyy-MM-dd'),
+                                        "status": punchMarkingTableData?.find((e) => v.sect_id === e.sect_id)?.salary_status ?? 0
+                                    }
+
+                                }),
+                            }
+                        })
+                        setDeptList(findDept)
+                        setOpenBkDrop(false)
+                        succesNofity(message)
+                    }
+                }
             } else {
-                console.log("error");
                 errorNofity(message)
                 setOpenBkDrop(false)
             }
         } else {
             errorNofity(message)
         }
-    }, [value])
+    }, [em_no])
 
 
     const exportDataClick = useCallback(() => {
@@ -349,6 +481,7 @@ const MonthlySalaryProcess = () => {
 
     return (
         <CustomLayout title="Monthly Salary Process" displayClose={true} >
+            <ToastContainer />
             <CustomBackDrop open={openBkDrop} text="!!! Please wait...Monthly Payroll Processing.... Do not Refesh or Reload the Browser !!!" />
             <Box sx={{ width: '100%', }}>
                 <Box sx={{ display: 'flex', py: 0.5, width: '100%', }}>
@@ -454,12 +587,25 @@ const MonthlySalaryProcess = () => {
                                                                     size="sm"
                                                                     variant="outlined"
                                                                 >Update Attendance</Chip> */}
-                                                            <Chip
+                                                            {
+                                                                e.status === 1 ? <Chip
+                                                                    color="danger"
+                                                                    onClick={() => deleteAttendanceMarkingProcess(row.dept_id, e.sect_id, e.updated)}
+                                                                    size="sm"
+                                                                    variant="outlined"
+                                                                >Delete Process</Chip> : <Chip
+                                                                    color="neutral"
+                                                                    // onClick={() => deleteAttendanceMarkingProcess(row.dept_id, e.sect_id, e.updated)}
+                                                                    size="sm"
+                                                                    variant="outlined"
+                                                                >Delete Process</Chip>
+                                                            }
+                                                            {/* <Chip
                                                                 color="danger"
                                                                 onClick={() => deleteAttendanceMarkingProcess(row.dept_id, e.sect_id, e.updated)}
                                                                 size="sm"
                                                                 variant="outlined"
-                                                            >Delete Process</Chip>
+                                                            >Delete Process</Chip> */}
                                                         </Box>
                                                     </td>
                                                 </tr>
